fix cert renew incompat for tedge-p11-server 1.6.1
If tedge-p11-server is too old, use previous method to renew only
existing certificates.

With tedge-p11-server 1.6.1, we could only renew existing certificates.
In later versions, where an ability to export the PEM of public keys was
added and we could use this to generate CSRs without existing
certificate, this new method replaced the previous approach, but this
introduced the incompatibility (since tedge-p11-server 1.6.1 didn't
support it). So now we revert to the older method if server is too old.

Signed-off-by: Marcel Guzik <marcel.guzik@cumulocity.com>